Update download script
