+++ extension-manifest-v3/package.json
@@ -7,7 +7,7 @@
     "postinstall": "npm run download-engines",
     "build": "node scripts/build.js",
     "start": "npm run build -- --watch",
-    "download-engines": "node scripts/download-engines.js",
+    "download-engines": "node scripts/download-engines/index.js",
     "licenses": "node ../tools/licenses.js",
     "lint": "eslint src/",
     "test": "npm run lint",
